25MinMAC v1
Quick script I threw together because £25 a day for wifi is extortion. This script takes advantage of the fact that the paywall offers a 30 minute trial to devices with unrecognised MAC addresses. By randomising the MAC address every 25 minutes the user is able to use the trial for as long as needed (having to click through the splash screen each time)